Move streamAnswer.ts to langgraph.
As it's tightly bound to langgraph.
